requirements.txt
numpy
Pillow
torch
torchvision
gradio